requirements.txt
streamlit
folium
streamlit-folium
pillow
reportlab